backend/requirements.txt
livekit-agents[openai,silero,cartesia]==1.2.18
python-dotenv==1.2.1
